refactor(map-storage): extract isModifyAreaMessageOnlyClaim function for personal area claim/revoke logic
